# docs/swagger.yaml
components:
  schemas:
...
    github_com_leseb_openresponses-gw_pkg_core_schema.CreatePromptRequest:
      properties:
        description:
          type: string
        metadata:
          type: object
        name:
          description: Required
          type: string
        template:
          description: Required
          type: string
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.DeletePromptResponse:
      properties:
        deleted:
          description: Always true
          type: boolean
        id:
          description: Prompt ID
          type: string
        object:
          description: Always "prompt.deleted"
          type: string
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.ListPromptsResponse:
      properties:
        data:
          description: Array of prompts
          items:
            $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Prompt'
          type: array
          uniqueItems: false
        first_id:
          description: ID of first item
          type: string
        has_more:
          description: Whether there are more results
          type: boolean
        last_id:
          description: ID of last item
          type: string
        object:
          description: Always "list"
          type: string
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.Prompt:
      properties:
        created_at:
          description: Unix timestamp
          type: integer
        description:
          type: string
        id:
          description: 'Format: "prompt_{uuid}"'
          type: string
        is_default:
          description: Whether this is the default version
          type: boolean
        metadata:
          type: object
        name:
          description: Human-readable name
          type: string
        object:
          description: Always "prompt"
          type: string
        template:
          description: Template with {{variables}}
          type: string
        updated_at:
          description: Unix timestamp
          type: integer
        variables:
          description: List of variable names
          items:
            type: string
          type: array
          uniqueItems: false
        version:
          description: Version number, starts at 1
          type: integer
      type: object
...
    github_com_leseb_openresponses-gw_pkg_core_schema.UpdatePromptRequest:
      properties:
        description:
          type: string
        metadata:
          type: object
        name:
          type: string
        set_as_default:
          description: Default true
          type: boolean
        template:
          type: string
        version:
          description: 'Required: current version being updated'
          type: integer
      type: object
...
info:
  contact:
    name: Open Responses Gateway
    url: https://github.com/leseb/openresponses-gw
  description: |-
    100% Open Responses Specification Compliant Gateway.
    Based on: https://github.com/openresponses/openresponses

    This gateway provides:
    - **Core API**: Full Open Responses spec compliance (POST /v1/responses)
    - **Extended APIs**: Conversations, Prompts, Files, Vector Stores, Connectors
    - **Dual Mode**: Standalone HTTP server or Envoy ExtProc integration

    Streaming: All 24 event types from Open Responses spec
    Request Echo: All request parameters returned in response
    Multimodal: Support for text, images, files, video
  title: Open Responses Gateway API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/prompts:
    get:
      parameters:
      - description: Cursor for pagination
        in: query
        name: after
        schema:
          type: string
      - description: Cursor for pagination (backwards)
        in: query
        name: before
        schema:
          type: string
      - description: Number of items (1-100, default 50)
        in: query
        name: limit
        schema:
          type: integer
      - description: 'Sort order: asc or desc (default desc)'
        in: query
        name: order
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListPromptsResponse'
          description: OK
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: List prompts
      tags:
      - Prompts
    post:
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.CreatePromptRequest'
                description: Create prompt request
                summary: request
        description: Create prompt request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Prompt'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Internal Server Error
      summary: Create prompt
      tags:
      - Prompts
  /v1/prompts/{id}:
    delete:
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.DeletePromptResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Delete prompt
      tags:
      - Prompts
    get:
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        schema:
          type: string
      - description: Specific version number
        in: query
        name: version
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Prompt'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: Get prompt
      tags:
      - Prompts
    put:
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.UpdatePromptRequest'
                description: Update prompt request
                summary: request
        description: Update prompt request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.Prompt'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Conflict
      summary: Update prompt
      tags:
      - Prompts
...
  /v1/prompts/{id}/versions:
    get:
      parameters:
      - description: Prompt ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/github_com_leseb_openresponses-gw_pkg_core_schema.ListPromptsResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: Not Found
      summary: List prompt versions
      tags:
      - Prompts
...
servers:
- description: Local development server
  url: http://localhost:8080
...